requirements.txt
aiohttp @ file:///C:/ci/aiohttp_1632749654317/work
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636523946399/work
chardet @ file:///C:/ci/chardet_1607706937985/work
charset-normalizer==2.0.7
click==8.0.3
cryptography @ file:///C:/ci/cryptography_1633520531101/work
dateparser @ file:///tmp/build/80754af9/dateparser_1604012780813/work
easyimap==0.7.0
httplib2==0.20.2
idna==3.3
multidict @ file:///C:/ci/multidict_1607349747897/work
mysql-connector-python==8.0.27
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
protobuf==3.19.1
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.6
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
regex @ file:///C:/ci/regex_1628063427816/work
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzlocal @ file:///C:/ci/tzlocal_1608245422044/work
uritemplate==4.1.1
urllib3==1.26.7
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
yarl @ file:///C:/ci/yarl_1606940155993/work
